requirements.txt
asgiref
boto3
botocore
Django
django-cors-headers
django-debug-toolbar
django-extensions
django-storages
djangorestframework
djangorestframework_simplejwt
Faker
gunicorn
jmespath
MarkupSafe
psycopg2-binary
pydot
PyJWT
pyparsing
python-dateutil
python-decouple
s3transfer
six
sqlparse
types-PyYAML
typing_extensions
tzdata
urllib3
Werkzeug
whitenoise
Pillow
